fix: Resolve metronome-BGM timing synchronization issues
- Fix BPM precision loss by changing Song.bpm from Int to Double throughout codebase
- Preserve precise BPM values (e.g. 165.55) instead of rounding to integers
- Resolve timing reference mismatch between BGM deviceCurrentTime and metronome CFAbsoluteTime
- Implement immediate metronome start with delayed BGM scheduling for perfect sync
- Update ServerSongDownloader, InputManager, and MetronomeComponent for Double BPM support
- Apply SwiftLint fixes for critical line length and closure parameter violations

Note: File/function length violations remain and require future refactoring

+++ Virgo/models/DrumTrack.swift
@@ -49,8 +49,8 @@ final class Chart {
     var artist: String {
         song?.artist ?? "Unknown Artist"
     }
-    var bpm: Int {
-        song?.bpm ?? 120
+    var bpm: Double {
+        song?.bpm ?? 120.0
     }
     var duration: String {
         song?.duration ?? "0:00"
@@ -84,7 +84,7 @@ final class Chart {
 final class Song {
     var title: String
     var artist: String
-    var bpm: Int
+    var bpm: Double
     var duration: String
     var genre: String
     private var _timeSignature: TimeSignature?
@@ -140,7 +140,7 @@ final class Song {
     init(
         title: String,
         artist: String,
-        bpm: Int,
+        bpm: Double,
         duration: String,
         genre: String,
         timeSignature: TimeSignature = .fourFour,
@@ -286,7 +286,7 @@ struct DrumTrack: Equatable {
     // Forward all properties to the chart and its song
     var title: String { chart.title }
     var artist: String { chart.artist }
-    var bpm: Int { chart.bpm }
+    var bpm: Double { chart.bpm }
     var duration: String { chart.duration }
     var genre: String { chart.genre }
     var difficulty: Difficulty { chart.difficulty }
